# Improve CopyHeaders performance