I was doing some benchmarks of code using utils.CopyHeaders, and while there isn't much to improve here, by knowing that `http.Headers` is really just a `map[string][]string`, we can append to the string slice directly.  Doesn't save any allocs, but does save about 20% in CPU time in my tests:

```
benchmark                  old ns/op     new ns/op     delta
BenchmarkCopyHeaders-4     1273          1016          -20.19%
```